Melhorias diversas.

+++ app/src/main/java/dev/ferrazpedro/livebooks2/data/model/LivroResposta.kt
@@ -2,7 +2,7 @@ package dev.ferrazpedro.livebooks2.data.model
 
 class LivroResposta(
     val title: String?,
-    val price: Double?,
+    val price: Float?,
     val writer: String?,
     val thumbnailHd: String?,
     val date: String?
