+++ app/src/main/java/dev/ferrazpedro/livebooks2/domain/entities/Livros.kt
@@ -4,7 +4,7 @@ import java.util.*
 
 class Livros(
     val titulo: String,
-    val preco: Double,
+    val preco: Float,
     val escritor: String,
     val caminhoImagem: String,
     val data: Date
